+++ examples/workspaces/fsharp-example/Program.fs
@@ -1,0 +1,5 @@
+// A simple F# console application
+printfn "Hello from F# in CodingBooth!"
+
+let sum = [1..10] |> List.sum
+printfn "Sum of 1..10 = %d" sum